BUGFIX: Use std::sqrtf for float
WHY: Be consistent when calling library functions.
HOW: Call the appropriate sqrt that fits the bill.

+++ common/src/main/datastructures.cpp
@@ -213,7 +213,7 @@ float MagSquared(tup const &Tup)
 }
 
 //------------------------------------------------------------------------------
-float Mag(tup const &Tup) { return (std::sqrt(MagSquared(Tup))); }
+float Mag(tup const &Tup) { return (std::sqrtf(MagSquared(Tup))); }
 
 //------------------------------------------------------------------------------
 tup Negate(tup const &Tup)
